questions-generator/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain
langchain-core
langchain-community
nltk
scikit-learn
rank-bm25
chardet
requests
